requirements.txt
# Dependencies for bot + scanner
python-telegram-bot
python-dotenv
requests